test(config): split loader_test.go into focused test files
loader_test.go held 12 test functions across four concerns. Split by
concern: defaults, format detection, file loading, and URL loading with
fallback. loader_test.go keeps only the testConfig fixture and the
containsString helper, which the file and URL tests share.

Pure move. No test logic, assertion, or name changed. Closes #47.
